requirements.txt
flask
flask-socketio
python-socketio
torch
torchvision
Pillow
numpy
opencv-python
pyyaml
scipy